chore: extract to same func
